Enhance BilgeBlog.Application and WebApi by adding FluentValidation for request validation, updating CreatePostCommand and UpdatePostCommand to include CategoryId and Tags properties, and modifying related command handlers to manage categories and tags. Update GetAllPostsQuery, GetPostCommentsQuery, and GetPostLikesQuery to return paged results. Adjust mapping profiles and request models accordingly for improved data handling.

+++ Core/BilgeBlog.Application/Handlers/CategoryHandlers/Modify/DeleteCategoryCommandHandler.cs
@@ -1,4 +1,5 @@
 using BilgeBlog.Application.DTOs.CategoryDtos.Commands;
+using BilgeBlog.Application.Exceptions;
 using BilgeBlog.Contract.Abstract;
 using MediatR;
 
@@ -15,6 +16,10 @@ namespace BilgeBlog.Application.Handlers.CategoryHandlers.Modify
 
         public async Task<bool> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
+            var category = await _categoryRepository.GetByIdAsync(request.Id);
+            if (category == null)
+                throw new NotFoundException("Category", request.Id);
+
             return await _categoryRepository.RemoveAsync(request.Id);
         }
     }
